site_builder/requirements.txt
feedgen==1.0
click==8.1
pyyaml==6.0.1
markdown==3.5